@@ -16,9 +16,18 @@
   "bugs": {
     "url": "https://github.com/JohnstonCode/svn-scm/issues"
   },
-  "categories": ["Other", "SCM Providers"],
-  "keywords": ["multi-root ready", "scm", "svn"],
-  "activationEvents": ["*"],
+  "categories": [
+    "Other",
+    "SCM Providers"
+  ],
+  "keywords": [
+    "multi-root ready",
+    "scm",
+    "svn"
+  ],
+  "activationEvents": [
+    "*"
+  ],
   "main": "./out/extension",
   "scripts": {
     "vscode:prepublish": "tsc -p ./",
@@ -31,12 +40,12 @@
     "jschardet": "^1.6.0"
   },
   "devDependencies": {
-    "typescript": "^2.5.2",
-    "vscode": "^1.1.5",
+    "typescript": "^2.6.1",
+    "vscode": "^1.1.6",
     "mocha": "^3.5.0",
     "eslint": "^4.6.1",
-    "@types/node": "^7.0.0",
-    "@types/mocha": "^2.2.42"
+    "@types/mocha": "^2.2.42",
+    "@types/node": "^7.0.43"
   },
   "contributes": {
     "commands": [
@@ -135,33 +144,41 @@
           "default": true
         },
         "svn.path": {
-          "type": ["string", "null"],
+          "type": [
+            "string",
+            "null"
+          ],
           "description": "Path to the svn executable",
           "default": null,
           "isExecutable": true
         },
         "svn.diff.withHead": {
           "type": "boolean",
-          "description":
-            "Show diff changes using latest revision in the repository. Set false to use latest revision in local folder",
+          "description": "Show diff changes using latest revision in the repository. Set false to use latest revision in local folder",
           "default": true
         },
         "svn.layout.trunk": {
-          "type": ["string", "null"],
-          "description":
-            "Relative path for 'trunk' in SVN URL, 'null' to disable. (Ex.: 'trunk', 'main')",
+          "type": [
+            "string",
+            "null"
+          ],
+          "description": "Relative path for 'trunk' in SVN URL, 'null' to disable. (Ex.: 'trunk', 'main')",
           "default": "trunk"
         },
         "svn.layout.branches": {
-          "type": ["string", "null"],
-          "description":
-            "Relative path for 'branches' in SVN URL, 'null' to disable. (Ex.: 'branches', 'versions')",
+          "type": [
+            "string",
+            "null"
+          ],
+          "description": "Relative path for 'branches' in SVN URL, 'null' to disable. (Ex.: 'branches', 'versions')",
           "default": "branches"
         },
         "svn.layout.tags": {
-          "type": ["string", "null"],
-          "description":
-            "Relative path for 'tags' in SVN URL, 'null' to disable. (Ex.: 'tags', 'stamps')",
+          "type": [
+            "string",
+            "null"
+          ],
+          "description": "Relative path for 'tags' in SVN URL, 'null' to disable. (Ex.: 'tags', 'stamps')",
           "default": "tags"
         },
         "svn.multipleFolders.enabled": {
