@@ -1,23 +1,21 @@
-/* global suite, test */
-
 //
 // Note: This example test is leveraging the Mocha test framework.
 // Please refer to their documentation on https://mochajs.org/ for help.
 //
 
 // The module 'assert' provides assertion methods from node
-var assert = require('assert');
+import * as assert from 'assert';
 
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
-var vscode = require('vscode');
-var myExtension = require('../extension');
+import * as vscode from 'vscode';
+import * as myExtension from '../extension';
 
 // Defines a Mocha test suite to group tests of similar kind together
-suite("Extension Tests", function() {
+suite("Extension Tests", () => {
 
     // Defines a Mocha unit test
-    test("Something 1", function() {
+    test("Something 1", () => {
         assert.equal(-1, [1, 2, 3].indexOf(5));
         assert.equal(-1, [1, 2, 3].indexOf(0));
     });
